+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+- 2019-12-02: Fixed multiple examples by updating "cost" to "goal", and other 
+              such changes. Fixed exampleMocoTrack by ignoring tendon 
+              compliance.
+
+- 2019-12-02: Add a MATLAB MocoTrajectory plotting utility, 
+              osimMocoTrajectoryReport.m. This has the same functionality as 
+              report.py in Moco's python package.
+
 - 2019-11-28: Brian Umberger contributed a Matlab version of example2DWalking,
               which contains a 2-D prediction of walking.